* REFA - renamed the init.py to serverconfig.py, and moved all configuration functions to it - separation of logic * FIX - the "list" CLI option works once again * FEAT - skipping of directory writability check, thus allowing to work under Windows
